[flatbuffers] Fix to working with vcpkg gRPC
The last flatbuffers release working with gRPC 1.14 but not with current
version. Because vcpkg only support the lastest version, flatbuffers+gRPC
use case is now not working with vcpkg.

This patch fix flatbuffers working with the latest gRPC.
When the next flatbuffers version is released with latest gRPC support,
this patch can be removed.